feat(wa-sqlite): split node and browser in package
